fix: update expense delete method, migrate amount to BigDecimal, fix precision bug

+++ src/main/java/com/expenseTracker/expensetracker/dto/CreateBudgetList.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -14,7 +15,7 @@ public class CreateBudgetList {
     @NotNull
     private String name;
     @NotNull
-    private Double budget;
+    private BigDecimal budget;
     @NotNull
     private LocalDate startDate;
     @NotNull
